Update index.php
Added an ISP check, probably a better way of doing it but it works fine in my own testing.

Removed the 'this is a beta, bla bla bla' text.

Removed the keywords meta tag. It's not really important.

Added a variable so you can define the description meta tag, may be useful. Leaving it blank is fine, too.